// frontend/libs/shared/data-access/src/lib/generated/models/inventory-balance-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { InventoryBalanceLine } from '../models/inventory-balance-line';
export interface InventoryBalanceResponse {
  balances: Array<InventoryBalanceLine>;
  location_ref_id: string;
  location_type: 'warehouse' | 'vehicle';
  tenant_id: string;
}